+++ lib/stimulus_spec/matchers.rb
@@ -1,5 +1,11 @@
 # frozen_string_literal: true
 
+require "nokogiri"
+
+require_relative "matchers/have_stimulus_controller"
+require_relative "matchers/have_stimulus_action"
+require_relative "matchers/have_stimulus_target"
+
 module StimulusSpec
   module Matchers
   end